# back-end-prod/back-end-prod/swagger-static/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/listings": {
      "get": {
        "operationId": "ListingController_fetchListings",
        "parameters": [
          {
            "name": "pageNumber",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "pageSize",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "order",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortedBy",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "dealType",
            "required": false,
            "in": "query",
            "schema": {
              "type": "array",
              "items": {}
            }
          },
          {
            "name": "industryType",
            "required": false,
            "in": "query",
            "schema": {
              "type": "array",
              "items": {}
            }
          },
          {
            "name": "city",
            "required": false,
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "maxPrice",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "minPrice",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "lastCursor",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ListingOverview"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Listings"
        ]
      },
      "post": {
        "operationId": "ListingController_registerListing",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrUpdateListingCommand"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListingOverview"
                }
              }
            }
          }
        },
        "tags": [
          "Listings"
        ],
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
    "/listings/seller-qna/{id}": {
      "post": {
        "operationId": "ListingController_createSellerQnA",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SellerQnAModel"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SellerQnAModel"
                }
              }
            }
          }
        },
        "tags": [
          "Listings"
        ],
        "security": [
          {
            "JWT": []
          }
        ]
      },
      "put": {
        "operationId": "ListingController_updateSellerQnA",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SellerQnAModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SellerQnAModel"
                }
              }
            }
          }
        },
        "tags": [
          "Listings"
        ],
        "security": [
          {
            "JWT": []
          }
        ]
      },
      "get": {
        "operationId": "ListingController_getSellerQnAByListingId",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SellerQnAModel"
                }
              }
            }
          }
        },
        "tags": [
          "Listings"
        ],
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
    "/listings/financial/{listingId}": {
      "post": {
        "operationId": "ListingController_addFinancial",
        "parameters": [
          {
            "name": "listingId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrUpdateFinancialCommand"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListingOverview"
                }
              }
            }
          }
        },
        "tags": [
          "Listings"
        ],
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
    "/listings/mylistings": {
      "get": {
        "operationId": "ListingController_getMyListings",
        "parameters": [
          {
            "name": "pageNumber",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "pageSize",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Listings"
        ],
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
    "/listings/{listingId}": {
      "get": {
        "operationId": "ListingController_getListingById",
        "parameters": [
          {
            "name": "listingId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListingDetails"
                }
              }
            }
          }
        },
        "tags": [
          "Listings"
        ]
      }
    },
    "/listings/business/{listingid}": {
      "get": {
        "operationId": "ListingController_getListingBusinessById",
        "parameters": [
          {
            "name": "listingid",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListingModel"
                }
              }
            }
          }
        },
        "tags": [
          "Listings"
        ]
      }
    },
    "/listings/upload/images/{listingId}": {
      "post": {
        "operationId": "ListingController_uploadListingImages",
        "parameters": [
          {
            "name": "listingId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "files": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Listings"
        ],
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
    "/listings/upload/documents/{listingId}": {
      "post": {
        "operationId": "ListingController_uploadListingPdfs",
        "parameters": [
          {
            "name": "listingId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "files": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Listings"
        ],
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
    "/listings/latest/{number}": {
      "get": {
        "operationId": "ListingController_getLatestListings",
        "parameters": [
          {
            "name": "number",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Listings"
        ]
      }
    },
    "/listings/update/isSold": {
      "put": {
        "operationId": "ListingController_updateIsSold",
        "parameters": [
          {
            "name": "listingId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateIsSold"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListingOverview"
                }
              }
            }
          }
        },
        "tags": [
          "Listings"
        ],
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
    "/listings/update/listing/{listingId}": {
      "patch": {
        "operationId": "ListingController_updateListing",
        "parameters": [
          {
            "name": "listingId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrUpdateListingCommand"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListingOverview"
                }
              }
            }
          }
        },
        "tags": [
          "Listings"
        ],
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
...
    "/listings/listing/financial/{listingId}": {
      "get": {
        "operationId": "ListingController_getFinancialByListingId",
        "parameters": [
          {
            "name": "listingId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Listings"
        ]
      }
    },
...
  },
  "info": {
    "title": "API",
    "description": "API docs",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ListingOverview": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "image": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "location": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "country": {
            "type": "string"
          },
          "dealType": {
            "type": "string"
          },
          "industryType": {
            "type": "string"
          },
          "annualRevenue": {
            "type": "number"
          },
          "annualNetProfit": {
            "type": "number"
          },
          "price": {
            "type": "number"
          },
          "isStartup": {
            "type": "boolean"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "ownerId": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "image",
          "title",
          "location",
          "city",
          "country",
          "dealType",
          "industryType",
          "annualRevenue",
          "annualNetProfit",
          "price",
          "isStartup",
          "createdAt",
          "ownerId"
        ]
      },
...
    }
  }
}